Optimize kernel logic
Instead of running a minimal block and grid size, use as many threads
per block as the platform allows, and calculate the grid size based
on that. This also requires us to calculate the array indices using
the thread and block dimesnions and IDs, getting rid of the loop in
our kernel.

Co-authored-by: Christian Weigelt <weigelt@kth.se>