Update original kmeans code to use Texture Object API
The old way of using CUDA texture<> references has been deprecated
since CUDA Toolchain 12, and the Rodinia suite hasn't been adapted
to that change. This meant that the project refused to compile on
our v12.3 toolchain, so we had to make some changes to the original
code base to even get it to compile.

Co-authored-by: Christian Weigelt <weigelt@kth.se>